preferences activity added

+++ app/src/main/java/rafaxplayer/misseries/models/Serie.java
@@ -2,8 +2,6 @@ package rafaxplayer.misseries.models;
 
 import com.google.firebase.database.IgnoreExtraProperties;
 
-import java.util.HashMap;
-
 /**
  * Created by rafax on 10/01/2017.
  */
@@ -13,20 +11,20 @@ public class Serie {
     public String name;
     public String poster;
     public String temps;
-    public int novistos;
+    public long novistos;
 
     public Serie() {
     }
 
     public Serie(String code,String name, String poster,String temps) {
-        this.code=code;
+        this.code = code;
         this.name = name;
         this.poster = poster;
-        this.temps=temps;
+        this.temps = temps;
 
     }
-    public void setnovistos(int novistos){
-        this.novistos=novistos;
+    public void setnovistos(long novistos){
+        this.novistos = novistos;
     }
